fix: OG 이미지 생성 TSX를 JS로 변환하여 빌드 오류 해결(#618)
