+++ vercel.json
@@ -1,8 +1,3 @@
 {
-  "functions": {
-    "api/og.tsx": {
-      "runtime": "edge"
-    }
-  },
   "rewrites": [{ "source": "/(.*)", "destination": "/" }]
 }
